requirements.txt
matplotlib==3.9.2
mlflow==2.16.2
mlflow_skinny==2.16.2
numpy==2.1.1
pandas==2.2.3
Pillow==10.4.0
scikit_learn==1.5.2
seaborn==0.13.2
timm==1.0.9
torch==2.4.1
torcheval==0.0.7
torchvision==0.19.1
